requirements.txt
fastapi
python-dotenv
uvicorn[standard]
requests